Moved around displays to create an abstract time display
